masters/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2.1
django-autoslug==1.9.9
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-redis
dotenv==0.9.9
drf-yasg==1.21.10
elasticsearch
django-elasticsearch-dsl
inflection==0.5.1
kombu==5.5.3
packaging==25.0
pillow==11.2.1
progressbar2==4.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
